PKG -- [fcl] deprecate fcl.verifyUserSignatures
