+++ packages/fcl/src/app-utils/index.js
@@ -1,1 +1,2 @@
 export {verifyAccountProof} from "./verify-account-proof.js"
+export {verifyUserSignatures} from "./verify-user-signatures.js"
